requirements.txt
openai==0.28
streamlit
pyaudio
python-dotenv